корректировки изменение типов данных с int на double вывод данных на экран в таблице

+++ AnimalsWPF/Birds.cs
@@ -15,17 +15,17 @@ namespace AnimalsWPF
         /// <summary>
         /// Возраст
         /// </summary>
-        public int Age { get; set; }
+        public double Age { get; set; }
         /// <summary>
         /// Вес
         /// </summary>
-        public int Weight { get; set; }
+        public double Weight { get; set; }
         /// <summary>
         /// Страна происхождения
         /// </summary>
         public string CountryFrom { get; set; }
 
-        public Birds(string name, int age, int weight, string countryFrom)
+        public Birds(string name, double age, double weight, string countryFrom)
         {
             this.Name = name;
             this.Age = age;
