+++ AnimalsWPF/Mammals.cs
@@ -16,17 +16,17 @@ namespace AnimalsWPF
         /// <summary>
         /// Возраст
         /// </summary>
-        public int Age { get; set; }
+        public double Age { get; set; }
         /// <summary>
         /// Вес
         /// </summary>
-        public int Weight { get; set; }
+        public double Weight { get; set; }
         /// <summary>
         /// Страна происхождения
         /// </summary>
         public string CountryFrom { get; set; }
 
-        public Mammals(string name, int age, int weight, string countryFrom)
+        public Mammals(string name, double age, double weight, string countryFrom)
         {
             this.Name = name;
             this.Age = age;
